requirements.txt
fastapi
uvicorn
pandas
opencv-python
Jinja2==3.0.3
fastapi-sqlalchemy
python-dotenv
flask
pillow
tensorflow
python-multipart
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
starlette==0.36.3
Werkzeug==3.0.1

